+++ src/main/java/com/mow/entity/Riders.java
@@ -29,7 +29,7 @@ public class Riders {
 	double rating;
 
 	@Column(name = "total_rating", nullable = false)
-	double totalRatings;
+	int totalRatings;
 	
 	@Column(name = "approved", nullable = false)
 	boolean approved;
